requirements.txt
Django==2.2.16
djangorestframework==3.12.2
drf-yasg==1.20.0
